pyproject.toml
[project]
name = "context-use"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-search-results>=2.4.2",
    "openai-agents>=0.3.2",
    "python-dotenv>=1.1.1",
]
